Add files via upload
Project Overview:
The main goal of this project was to create an easy-to-use dashboard that allows for real-time tracking and detailed analysis of sales data for Madhav Store. Here’s what I did:

1. Data Integration: Connected and combined multiple data sources for a complete dataset.
2. Data Manipulation: Used advanced calculations to clean and adjust the data for accurate insights.
3. Dynamic Filters: Added user-driven parameters so users can customize the data view based on their needs.
4. Drill-Down Insights: Created drill-down options for more detailed analysis.
5. Interactive Dashboard: Used filters and slicers to make the dashboard user-friendly and interactive.
6. Custom Visuals: Applied different visualizations to display the data in an engaging way.

This project improved my skills in data integration, manipulation, and visualization, and the final dashboard helps Madhav Store make informed, data-driven decisions.